packages and files renamed for a more consistent naming scheme
