+++ demo/src/main/scala/littleMonadDemo/ch06FunctionsAreValues/Example06.scala
@@ -1,8 +1,8 @@
-package littleMonadDemo.ch06ListOfFunctions
+package littleMonadDemo.ch06FunctionsAreValues
 
 import scala.util.chaining._
 
-object ListOfFunctions extends util.App {
+object Example06 extends util.App {
 
   val plus1: Function1[Int, Int] = x => x + 1
   def plus2(x: Int): Int         = x + 2
