--- demo/src/main/scala/littleMonadDemo/ch07ImportanceOfFunction1/ImportanceOfFunction1.scala
@@ -1,8 +1,0 @@
-package littleMonadDemo.ch07ImportanceOfFunction1
-
-import scala.util.chaining._
-
-object ImportanceOfFunction1 extends util.App {
-
-  "The Importance of Function1" pipe println
-}
