--- demo/src/main/scala/littleMonadDemo/ch07ImportanceOfFunction1/README.md
@@ -1,4 +1,0 @@
-# Chapter 07:
-# The Importance of *Function1*
-
-TODO
